requirements.txt
google-cloud-storage==2.14.0
Flask==3.0.0
gunicorn==20.1.0
Werkzeug==3.0.1
python-dotenv==1.0.0


